[Refactor] enum 제네릭 사용

+++ Assets/Project/Develop/NSJ/Script/UI/GameCanvas.cs
@@ -12,6 +12,6 @@ public class GameCanvas : BaseCanvas
 
     void OnEnable()
     {
-        ChangePanel((int)Panel.Game);
+        ChangePanel(Panel.Game);
     }
 }
