+++ Assets/Project/Develop/NSJ/Script/UI/TitleBox.cs
@@ -18,7 +18,7 @@ public class TitleBox : BaseBox
 
     private void SubscribesEvent()
     {
-        _start.onClick.AddListener(() => Panel.ChangeBox((int)TitlePanel.Box.Start));
+        _start.onClick.AddListener(() => Panel.ChangeBox(TitlePanel.Box.Start));
         _exit.onClick.AddListener(ExitGame);
     }
     private void ExitGame()
